// frontend/src/shared/api/generated/models/eventCategoryTypeResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * CodigoActivo API
 * Backend API for CodigoActivo (cookie session auth + CSRF).
 * OpenAPI spec version: v1
 */

export interface EventCategoryTypeResponse {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  color?: string | null;
}